Teste de Orcamento e ServicoConcluido
Funcionando!
+ajustes nas classes Orcamento e ServicoConcluido e PKs

+++ src/br/com/facility/bo/OrcamentoBO.java
@@ -37,7 +37,7 @@ public class OrcamentoBO {
 		
 	}
 	
-	public Orcamento consultar(int codigo, Negociacao n){
+	public Orcamento consultar(long codigo, Negociacao n){
 		
 		OrcamentoPK oPK = new OrcamentoPK();
 		oPK.setCodigo(codigo);
